backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
qdrant-client==1.10.1
matplotlib
python-multipart
scikit-learn
jinja2
